ref [Bug][NII Redmine#38015]機関ストレージ(アドオン方式)で接続した「Amazon S3 Compatible Storage for Institution」に対してクオータ制御が適切に動作しない: Check max creator quota
