+++ waterbutler/server/api/v1/provider/create.py
@@ -94,7 +94,7 @@ class CreateMixin:
             self.target_path = self.path
 
         # verify quota if it is osfstorage
-        if self.provider.NAME == 'osfstorage':
+        if self.provider.NAME == 'osfstorage' or self.provider.NAME == 's3compatinstitutions':
             file_size = int(self.request.headers.get('Content-Length'))
             quota = await self.provider.get_quota()
             if quota['used'] + file_size > quota['max']:
